Changed Datatype of 'product' varible from int to long

+++ main.java
@@ -4,7 +4,7 @@ public class main {
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter Number");
         int n =sc.nextInt();
-        int product =1;
+        long product =1;
         for(int i = 1; i <= n;  i ++){
             product = product * i; 
         }
